requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
requests==2.32.3
google-auth==2.35.0
google-api-python-client==2.154.0
gunicorn==23.0.0
Pillow==11.2.1
python-dotenv==1.0.1
Werkzeug==3.0.6
anthropic==0.97.0
beautifulsoup4==4.14.3
lxml==6.1.0
